Skeleton of classes reference view

+++ Classes/UIImage+PDF/PDFView.m
@@ -109,7 +109,7 @@
 	{
 		CGPDFDocumentRef pdf = CGPDFDocumentCreateWithURL( (__bridge CFURLRef) resourceURL );
 		
-		pageCount = CGPDFDocumentGetNumberOfPages( pdf );
+		pageCount = (int)CGPDFDocumentGetNumberOfPages( pdf );
 		
 		CGPDFDocumentRelease( pdf );
 	}
